Update paired devices labels
